LayoutTests/imported/w3c/web-platform-tests/css/css-mixins/mixin-from-import-with-media-queries-expected.txt

This text should be green.

FAIL CSS Mixins: @import with media queries assert_equals: expected "rgb(0, 128, 0)" but got "rgb(255, 0, 0)"
